Refactor scripts to src dir. Rewrite __init__.py.
