+++ setup.py
@@ -5,5 +5,5 @@ setup(
     version="0.0.1",
     description="Wiki dump extraction for spaCy",
     url="https://www.github.com/explosion/wikid/",
-    packages=["scripts", "scripts.wiki"],
+    packages=["src", "src.wiki"],
 )
